requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==4.0.0
exceptiongroup==1.1.3
FastAPI-SQLAlchemy==0.2.1
greenlet==3.0.1
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.32.0.post1
typing_extensions==4.8.0
